lsp: extract i18n edit-building into I18nHandler
Pure move out of DiagnosticsHandler (48K and growing) ahead of the
i18n translation actions (#5283). No behavior change.

Also retargets the buildAddExtractEdit call sites that moved with it:
navigation.js's CodeActionHandler construction and 13 direct calls in
diagnostics.js's extract-edit unit tests, neither flagged in the task
brief but both required to keep those suites passing.
