+++ tools/tests/testFoamLSP.js
@@ -38,6 +38,7 @@ if ( require.main === module ) {
 //   node foam3/tools/tests/testFoamLSP.js hover,completion # comma- or space-separated
 var CATEGORIES = [
   'foamIndex', 'grammar', 'utilities', 'completion', 'hover', 'diagnostics',
+  'i18n',
   'navigation', 'java', 'jrl', 'editorFeatures', 'typeHierarchy', 'usageIndex',
   'callHierarchy', 'pomValidation', 'pomNavigation', 'mcp'
 ];
